pyproject.toml
[tool.poetry]
name = "storystudio"
version = "0.1.0"
description = "Based on generative ai to auto-generate story video"
authors = ["MIBlue119 <miblue119@gmail.com>"]
readme = "README.md"
packages = [{include = "storystudio"}]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.27.8"
semantic-kernel = "^0.3.5.dev0"
loguru = "^0.7.0"
azure-cognitiveservices-speech = "^1.31.0"
replicate = "^0.11.0"
pydantic-settings = "^2.0.2"
pydantic = "^2.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
